Refactored to put all validation inside form
